requirements.txt
datasets
einops
sentencepiece
huggingface-hub
nvitop
accelerate
diffusers
transformers==4.45.2
wandb
deepspeed
peft==0.18.1
imageio
numpy
opencv-python
omegaconf
pandas
pillow
protobuf
pyarrow
tensorboard
tokenizers
tqdm
trl==0.8.6
fire
matplotlib
rich